R CMD install now works
